static class var idCounter in ESEEvent
I wanted to start testing the most simple things and I realized
that when I want to create an Event, I have to decide for an id
number.
I thought it would be the best, if there was just an class variable
that gets incremented by each new Event.
May be we will have to change id int to id long.

This changement means that I also changed the Constructor of Event!

	modified:   app/models/ESEEvent.java